Periodic check-in

+++ src/boids/BoidArray.java
@@ -2,8 +2,8 @@ package boids;
 
 public record BoidArray(
         int n,
-        double[] x,
-        double[] y,
+        int[] x,
+        int[] y,
         int[] h,
         double tick
 ) {}
